Migrate Settings family to split-class layout
8 resources migrated:

  ChannelResource         173 -> 73 lines  (+ChannelForm 89, +ChannelTable 55)
  CurrencyResource        166 -> 65 lines  (+CurrencyForm 100, +CurrencyTable 52)
  LanguageResource        113 -> 64 lines  (+LanguageForm 60, +LanguageTable 43)
  ActivityResource        172 -> 60 lines  (+ActivityForm 47, +ActivityTable 80)
  StaffResource           224 -> 75 lines  (+StaffForm 152, +StaffTable 53)
  TaxClassResource        121 -> 65 lines  (+TaxClassForm 50, +TaxClassTable 50)
  TaxRateResource         131 -> 75 lines  (+TaxRateForm 65, +TaxRateTable 47)
  TaxZoneResource         368 -> 65 lines  (+TaxZoneForm 270, +TaxZoneTable 67)

TaxZone's private syncCountries/syncStates/syncPostcodes helpers move
to TaxZoneForm with it — they're tightly coupled to the form save
closures, no external consumers.

Tax cluster resources (TaxClass, TaxRate, TaxZone) remain in the
Filament\Clusters\Taxes cluster; cluster registration is unchanged.

ResourceExtensionTest rewritten to test the new split-class hook
pattern (configureForm/configureTable on {Resource}Form / {Resource}Table)
since every test target is now migrated. Legacy extendForm/extendTable
test cases on Resource::class removed — that contract is being deleted
in the wrapper-trait removal step.
