+++ packages/admin/src/Filament/Resources/ActivityResource/Tables/ActivityTable.php
@@ -1,0 +1,89 @@
+<?php
+
+namespace Lunar\Admin\Filament\Resources\ActivityResource\Tables;
+
+use Carbon\Carbon;
+use Filament\Forms\Components\DatePicker;
+use Filament\Tables\Columns\TextColumn;
+use Filament\Tables\Filters\Filter;
+use Filament\Tables\Filters\SelectFilter;
+use Filament\Tables\Table;
+use Illuminate\Database\Eloquent\Builder;
+use Lunar\Admin\Support\Concerns\CallsHooks;
+
+class ActivityTable
+{
+    use CallsHooks;
+
+    public static function configure(Table $table): Table
+    {
+        return self::callStaticLunarHook(
+            'configureTable',
+            $table
+                ->columns(static::getColumns())
+                ->filters([
+                    SelectFilter::make('event')
+                        ->label(__('lunarpanel::activity.table.event'))
+                        ->multiple()
+                        ->options([
+                            'created' => 'Created',
+                            'updated' => 'Updated',
+                            'deleted' => 'Deleted',
+                        ]),
+                    Filter::make('created_at')
+                        ->schema([
+                            DatePicker::make('logged_from')
+                                ->label(__('lunarpanel::activity.table.logged_from')),
+                            DatePicker::make('logged_until')
+                                ->label(__('lunarpanel::activity.table.logged_until')),
+                        ])
+                        ->query(function (Builder $query, array $data): Builder {
+                            return $query
+                                ->when(
+                                    $data['logged_from'],
+                                    fn (Builder $query, $date): Builder => $query->whereDate('created_at', '>=', $date),
+                                )
+                                ->when(
+                                    $data['logged_until'],
+                                    fn (Builder $query, $date): Builder => $query->whereDate('created_at', '<=', $date),
+                                );
+                        })
+                        ->indicateUsing(function (array $data): array {
+                            $indicators = [];
+
+                            if ($data['logged_from'] ?? null) {
+                                $indicators['logged_from'] = 'Created from '.Carbon::parse($data['logged_from'])->toFormattedDateString();
+                            }
+
+                            if ($data['logged_until'] ?? null) {
+                                $indicators['logged_until'] = 'Created until '.Carbon::parse($data['logged_until'])->toFormattedDateString();
+                            }
+
+                            return $indicators;
+                        }),
+                ])
+                ->toolbarActions([])
+                ->defaultSort('id', 'DESC'),
+        );
+    }
+
+    public static function getColumns(): array
+    {
+        return [
+            TextColumn::make('id')
+                ->label('ID')
+                ->sortable(),
+            TextColumn::make('subject_type')
+                ->label(__('lunarpanel::activity.table.subject'))
+                ->searchable(),
+            TextColumn::make('description')
+                ->label(__('lunarpanel::activity.table.description'))
+                ->searchable(),
+            TextColumn::make('log_name')
+                ->label(__('lunarpanel::activity.table.log')),
+            TextColumn::make('created_at')
+                ->label(__('lunarpanel::activity.table.logged_at'))
+                ->dateTime(),
+        ];
+    }
+}
