+++ packages/admin/src/Filament/Resources/LanguageResource/Tables/LanguageTable.php
@@ -1,0 +1,40 @@
+<?php
+
+namespace Lunar\Admin\Filament\Resources\LanguageResource\Tables;
+
+use Awcodes\BadgeableColumn\Components\Badge;
+use Awcodes\BadgeableColumn\Components\BadgeableColumn;
+use Filament\Tables\Columns\TextColumn;
+use Filament\Tables\Table;
+use Illuminate\Database\Eloquent\Model;
+use Lunar\Admin\Support\Concerns\CallsHooks;
+
+class LanguageTable
+{
+    use CallsHooks;
+
+    public static function configure(Table $table): Table
+    {
+        return self::callStaticLunarHook(
+            'configureTable',
+            $table->columns(static::getColumns()),
+        );
+    }
+
+    public static function getColumns(): array
+    {
+        return [
+            BadgeableColumn::make('name')
+                ->separator('')
+                ->suffixBadges([
+                    Badge::make('default')
+                        ->label(__('lunarpanel::language.table.default.label'))
+                        ->color('gray')
+                        ->visible(fn (Model $record) => $record->default),
+                ])
+                ->label(__('lunarpanel::language.table.name.label')),
+            TextColumn::make('code')
+                ->label(__('lunarpanel::language.table.code.label')),
+        ];
+    }
+}
