+++ packages/admin/src/Filament/Resources/TaxZoneResource/Tables/TaxZoneTable.php
@@ -1,0 +1,60 @@
+<?php
+
+namespace Lunar\Admin\Filament\Resources\TaxZoneResource\Tables;
+
+use Awcodes\BadgeableColumn\Components\Badge;
+use Awcodes\BadgeableColumn\Components\BadgeableColumn;
+use Filament\Actions\BulkActionGroup;
+use Filament\Actions\DeleteBulkAction;
+use Filament\Actions\EditAction;
+use Filament\Tables\Columns\IconColumn;
+use Filament\Tables\Columns\TextColumn;
+use Filament\Tables\Table;
+use Illuminate\Database\Eloquent\Model;
+use Lunar\Admin\Support\Concerns\CallsHooks;
+
+class TaxZoneTable
+{
+    use CallsHooks;
+
+    public static function configure(Table $table): Table
+    {
+        return self::callStaticLunarHook(
+            'configureTable',
+            $table
+                ->columns(static::getColumns())
+                ->filters([])
+                ->recordActions([
+                    EditAction::make(),
+                ])
+                ->toolbarActions([
+                    BulkActionGroup::make([
+                        DeleteBulkAction::make(),
+                    ]),
+                ]),
+        );
+    }
+
+    public static function getColumns(): array
+    {
+        return [
+            BadgeableColumn::make('name')
+                ->separator('')
+                ->suffixBadges([
+                    Badge::make('default')
+                        ->label(__('lunarpanel::taxzone.table.default.label'))
+                        ->color('gray')
+                        ->visible(fn (Model $record) => $record->default),
+                ])
+                ->label(__('lunarpanel::taxzone.table.name.label')),
+            TextColumn::make('zone_type')
+                ->label(__('lunarpanel::taxzone.table.zone_type.label'))
+                ->formatStateUsing(
+                    fn ($state) => __("lunarpanel::taxzone.form.zone_type.options.{$state}")
+                ),
+            IconColumn::make('active')
+                ->boolean()
+                ->label(__('lunarpanel::taxzone.table.active.label')),
+        ];
+    }
+}
