+++ tests/admin/Unit/Support/Extending/ResourceExtensionTest.php
@@ -5,15 +5,21 @@ use Filament\Schemas\Schema;
 use Filament\Tables\Columns\TextColumn;
 use Filament\Tables\Table;
 use Livewire\Livewire;
-use Lunar\Admin\Filament\Resources\ActivityResource;
 use Lunar\Admin\Filament\Resources\ActivityResource\Pages\ListActivities;
+use Lunar\Admin\Filament\Resources\ActivityResource\Tables\ActivityTable;
 use Lunar\Admin\Filament\Resources\AttributeGroupResource\Pages\ListAttributeGroups;
 use Lunar\Admin\Filament\Resources\AttributeGroupResource\Tables\AttributeGroupTable;
 use Lunar\Admin\Filament\Resources\CurrencyResource;
+use Lunar\Admin\Filament\Resources\CurrencyResource\Pages\EditCurrency;
 use Lunar\Admin\Filament\Resources\CurrencyResource\Pages\ListCurrencies;
+use Lunar\Admin\Filament\Resources\CurrencyResource\Schemas\CurrencyForm;
+use Lunar\Admin\Filament\Resources\CurrencyResource\Tables\CurrencyTable;
 use Lunar\Admin\Filament\Resources\CustomerResource;
 use Lunar\Admin\Filament\Resources\LanguageResource;
+use Lunar\Admin\Filament\Resources\LanguageResource\Pages\EditLanguage;
 use Lunar\Admin\Filament\Resources\LanguageResource\Pages\ListLanguages;
+use Lunar\Admin\Filament\Resources\LanguageResource\Schemas\LanguageForm;
+use Lunar\Admin\Filament\Resources\LanguageResource\Tables\LanguageTable;
 use Lunar\Admin\Support\Extending\ResourceExtension;
 use Lunar\Admin\Support\Facades\LunarPanel;
 use Lunar\Tests\Admin\Feature\Filament\TestCase;
@@ -22,7 +28,7 @@ use Lunar\Tests\Admin\Stubs\Filament\TestCustomerAddressRelationManager;
 uses(TestCase::class)
     ->group('extending', 'extending.resources');
 
-it('can extend relationship managers', function () {
+it('can extend relationship managers via getRelations hook on the resource', function () {
     $class = new class extends ResourceExtension
     {
         public function getRelations(array $managers): array
@@ -41,32 +47,7 @@ it('can extend relationship managers', function () {
     expect($relations)->toContain(TestCustomerAddressRelationManager::class);
 });
 
-it('can extend table columns', function ($resource, $page) {
-    $class = new class extends ResourceExtension
-    {
-        public function extendTable(Table $table): Table
-        {
-            return $table->columns([
-                ...$table->getColumns(),
-                TextColumn::make('test_column'),
-            ]);
-        }
-    };
-
-    LunarPanel::extensions([
-        $resource => $class::class,
-    ]);
-
-    $this->asStaff();
-
-    Livewire::test($page)->assertTableColumnExists('test_column');
-})->with([
-    'ListCurrencies' => [CurrencyResource::class, ListCurrencies::class],
-    'ListLanguages' => [LanguageResource::class, ListLanguages::class],
-    'ListActivities' => [ActivityResource::class, ListActivities::class],
-]);
-
-it('can extend table columns on split-class tables (configureTable hook)', function () {
+it('can extend table columns via configureTable hook on the split-class table', function ($table, $page) {
     $class = new class extends ResourceExtension
     {
         public function configureTable(Table $table): Table
@@ -79,18 +60,23 @@ it('can extend table columns on split-class tables (configureTable hook)', funct
     };
 
     LunarPanel::extensions([
-        AttributeGroupTable::class => $class::class,
+        $table => $class::class,
     ]);
 
     $this->asStaff();
 
-    Livewire::test(ListAttributeGroups::class)->assertTableColumnExists('test_column');
-});
+    Livewire::test($page)->assertTableColumnExists('test_column');
+})->with([
+    'CurrencyTable' => [CurrencyTable::class, ListCurrencies::class],
+    'LanguageTable' => [LanguageTable::class, ListLanguages::class],
+    'ActivityTable' => [ActivityTable::class, ListActivities::class],
+    'AttributeGroupTable' => [AttributeGroupTable::class, ListAttributeGroups::class],
+]);
 
-it('can extend form schema', function ($resource, $page) {
+it('can extend form schema via configureForm hook on the split-class form', function ($resource, $form, $page) {
     $class = new class extends ResourceExtension
     {
-        public function extendForm(Schema $schema): Schema
+        public function configureForm(Schema $schema): Schema
         {
             $schema->components([
                 ...$schema->getComponents(true),
@@ -102,7 +88,7 @@ it('can extend form schema', function ($resource, $page) {
     };
 
     LunarPanel::extensions([
-        $resource => $class::class,
+        $form => $class::class,
     ]);
 
     $this->asStaff(admin: true);
@@ -113,6 +99,6 @@ it('can extend form schema', function ($resource, $page) {
         'record' => $model->getRouteKey(),
     ])->assertFormFieldExists('test_form_field');
 })->with([
-    'CurrencyResource' => [CurrencyResource::class, CurrencyResource\Pages\EditCurrency::class],
-    'LanguageResource' => [LanguageResource::class, LanguageResource\Pages\EditLanguage::class],
+    'CurrencyForm' => [CurrencyResource::class, CurrencyForm::class, EditCurrency::class],
+    'LanguageForm' => [LanguageResource::class, LanguageForm::class, EditLanguage::class],
 ]);
